Onboarding refactor into smaller components
